requirements.txt
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
chardet==5.2.0
click==8.1.8
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
load-dotenv==0.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg==3.2.4
psycopg2-binary==2.9.10
pypdf==5.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
rpds-py==0.23.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
